pydeploy/requirements.txt
pygit2==0.26.3
Django==2.0.1
docker_py==1.10.6
